faucet: move limiter to another file
