+++ cmd/faucet/faucet.go
@@ -49,7 +49,6 @@ import (
 	"github.com/ethereum/go-ethereum/log"
 	"github.com/ethereum/go-ethereum/params"
 	"github.com/gorilla/websocket"
-	lru "github.com/hashicorp/golang-lru"
 	"golang.org/x/time/rate"
 )
 
@@ -932,49 +931,4 @@ func getGenesis(genesisFlag string, goerliFlag bool, sepoliaFlag bool) (*core.Ge
 	default:
 		return nil, errors.New("no genesis flag provided")
 	}
-}
-
-type IPRateLimiter struct {
-	ips *lru.Cache
-	mu  *sync.RWMutex
-	r   rate.Limit
-	b   int
-}
-
-func NewIPRateLimiter(r rate.Limit, b int, size int) (*IPRateLimiter, error) {
-	cache, err := lru.New(size)
-	if err != nil {
-		return nil, err
-	}
-
-	i := &IPRateLimiter{
-		ips: cache,
-		mu:  &sync.RWMutex{},
-		r:   r,
-		b:   b,
-	}
-
-	return i, nil
-}
-
-func (i *IPRateLimiter) AddIP(ip string) *rate.Limiter {
-	i.mu.Lock()
-	defer i.mu.Unlock()
-
-	limiter := rate.NewLimiter(i.r, i.b)
-
-	i.ips.Add(ip, limiter)
-
-	return limiter
-}
-
-func (i *IPRateLimiter) GetLimiter(ip string) *rate.Limiter {
-	i.mu.Lock()
-	defer i.mu.Unlock()
-
-	if limiter, exists := i.ips.Get(ip); exists {
-		return limiter.(*rate.Limiter)
-	}
-
-	return i.AddIP(ip)
 }
